Move api/server/router/checkpoint to daemon/server/router/checkpoint
